Restructured components to make it more clear
