+++ client/src/components/description/Cover.jsx
