+++ client/src/components/description/Info.jsx
