+++ client/src/components/description/Snippet.jsx
