+++ client/src/components/gallery/Highlight.jsx
